Update source for using config in extracting acoustic features
