+++ disvoice/__init__.py
@@ -1,17 +1,14 @@
 # -*- coding: utf-8 -*-
-
-
 from .utils import *
 from .script_mananger import script_manager
 
-
 from . import (glottal, phonation, phonological, prosody, replearning, articulation)
-from .glottal.glottal import Glottal
-from .phonation.phonation import Phonation
-from .phonological.phonological import Phonological
-from .prosody.prosody import Prosody
-from .replearning.replearning import RepLearning
-from .articulation.articulation import Articulation
+from .glottal import Glottal
+from .phonation import Phonation
+from .phonological import Phonological
+from .prosody import Prosody
+from .replearning import RepLearning
+from .articulation import Articulation
