+++ disvoice/prosody/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from .prosody import Prosody
+# from ..prosody import Prosody
 from .prosody_functions import *
-
+from .Prosody import *
 
 __all__=["Prosody"]
